+++ CHANGELOG.rst
@@ -16,7 +16,7 @@ Changelog
 
 **Misc**
 
-- Bump minimum recommended transformers version to 5.0.
+- Add experimental support for transformers>=5.0. Unified Hugging Face checkpoint export for quantized checkpoints may not work for some models with transformers>=5.0 yet.
 
 0.43 (2026-04-09)
 ^^^^^^^^^^^^^^^^^